refactor: optimize aix grid component
